+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Changes
 - Generalize locale values for fault `invalid` and `unspecified`
+- Nest attribute error messages under `error` key within metadata
 
 ## [1.7.5] - 2025-09-10