+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### TODO:
 - Update exceptions with more info on how to fix the issue
+- Add trnasform option to attributes
 - Add option to output failure backtraces
 - Add durability (retries) to execution
  - N-retries (3 default)